[66] moved tip detection to command
